--- src/MEDS_transforms/configs/_preprocess.yaml
@@ -1,35 +1,0 @@
-defaults:
-  - _pipeline
-  - stage_configs:
-      - reshard_to_split
-      - filter_subjects
-      - add_time_derived_measurements
-      - count_code_occurrences
-      - filter_measurements
-      - fit_outlier_detection
-      - occlude_outliers
-      - fit_normalization
-      - fit_vocabulary_indices
-      # This stage is intentionally included as a stage config but not a stage so it is not executed in
-      # default pipelines but if someone adds it to their list of stages it can be run like normal.
-      - reorder_measurements
-  - _self_
-
-etl_metadata.pipeline_name: "preprocess"
-
-# Global pipeline parameters:
-# 1. Code modifiers will be used as adjoining parts of the `code` columns during group-bys and eventual
-#    tokenization.
-code_modifiers: ???
-
-# Pipeline Structure
-stages:
-  - filter_subjects
-  - add_time_derived_measurements
-  - preliminary_counts
-  - filter_measurements
-  - fit_outlier_detection
-  - occlude_outliers
-  - fit_normalization
-  - fit_vocabulary_indices
-  - normalization
