+++ src/MEDS_transforms/configs/stages/fit_vocabulary_indices.yaml
